Refactor client into new namespaces
